feat(bin): split main and experimental binaries
